+++ include/coli/graphics/context.h
@@ -3,28 +3,51 @@
 
 #include "coli/utility.h"
 
+namespace Coli::Graphics::Detail::inline OpenGL {
+    class WindowFactory;
+}
+
 namespace Coli::Graphics::inline OpenGL
 {
     /**
      * @brief The OpenGL context class
      *
-     * @warning This class blocks access from different threads
-     * @warning Only one existing instance of this class is allowed
+     * @warning This class blocks any access from different threads
+     * @warning Only one active instance of this class is allowed
      *
      * @note You must create this class via std::make_shared
      */
     class COLI_EXPORT Context final
     {
+        /// @throws std::logic_error
         [[noreturn]] static void fail_incorrect_thread();
+
+        /// @throws std::logic_error
         [[noreturn]] static void fail_already_exists();
+
+        /// @throws std::runtime_error
         [[noreturn]] static void fail_initialize_error();
+
+        /// @throws std::logic_error
         [[noreturn]] static void fail_not_ready();
+
+        /// @throws std::invalid_argument
+        [[noreturn]] static void fail_not_owner();
 
         class Tools;
 
-        friend class Window;
+        friend class Graphics::Detail::OpenGL::WindowFactory;
 
-        void register_window(GLFWwindow* handle) noexcept;
+        /// @throws std::runtime_error
+        static void load_opengl(GLFWwindow* handle);
+
+        /**
+         * @throws std::runtime_error
+         * @throws std::logic_error
+         */
+        void register_window(GLFWwindow* handle);
+
+        void unregister_window() noexcept;
 
     public:
         /**
@@ -40,7 +63,7 @@ namespace Coli::Graphics::inline OpenGL
          *
          * @param other The context to move from
          *
-         * @throw std::runtime_error If the creation thread do not match with the current one
+         * @throw std::logic_error If the creation thread do not match with the current one
          */
         Context(Context&& other) noexcept(false);
 
@@ -56,16 +79,32 @@ namespace Coli::Graphics::inline OpenGL
         /**
          * @brief Verify the creation thread and the current one match
          *
-         * @throw std::runtime_error If the threads do not match
+         * @throw std::logic_error If the threads do not match
          */
         void verify_thread() const;
 
         /**
          * @brief Verify the context is ready. It should be initialized and have a window
          *
-         * @throw std::runtime_error If the context are not ready
+         * @throw std::logic_error If the context are not ready
          */
         void verify_context() const;
+
+        /**
+         * @brief Checks does the context has a window
+         *
+         * @throw std::invalid_argument If calls on a not owning object
+         *
+         * @return True if the context has a window, false otherwise
+         */
+        [[nodiscard]] bool has_window() const;
+
+        /**
+         * @brief Checks the object owns the context
+         *
+         * @return True if this object is owning the context, false otherwise
+         */
+        [[nodiscard]] bool is_owning() const noexcept;
 
     private:
         std::thread::id myCreationThread;